Moved hk_driver fixture to local conftest.py
